Player Creation
Some changes were made to create temporary players.
For now each player gets a deck of cards.
A form will be created in one of the next commit to ask how many players will play and what are their info.
Next update will contain a method that check if a card can be placed according to the gamerules.