Placement conditions Creation
As said earlier, there are conditions to place a card. You must place it next to another card, or you can place it on top of a card with a lesser value (even if it is the same color).

Last condition that will be implemented is that the surface the cards occupies must be a 6x6 grid. So the board will reduce itself according to players' movements.