Checking Streak Length
If a player has a 4 cards streak (or 5 for two players), the game stops.
The streak can be a horizontal, vertical or diagonal one.

For now, this condition only stops the game, but it will trigger the end of the round in later commits.

The code that checks the length of streaks can also be used to break a tie.